Refactored screen editor namespace as designer namespace
Necessary to remain consistent with our action of the same name so these associated objects are easier to find and lookup.

Milestone: minor
